% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdplugr.R
\docType{package}
\name{rmdplugr-package}
\alias{rmdplugr}
\alias{rmdplugr-package}
\title{rmdplugr: Plugins for Rmarkdown Formats}
\description{
This package adds plugins that transform 'latex' templates
    used in Rmarkdown formats without relying on custom templates. The goal is 
    both to restore the
    style modifications introduced by the default behavior of the
    'rmarkdown' and 'pandoc' templates, such as
    smaller margins, paragraph spacing, and compact titles, and also
    to enable affiliations through the 'latex' 'auhthblk' package and
    marginal material using either 'scrlayer-page' or 'fancyhdr'.
}
\author{
\strong{Maintainer}: Johan Larsson \email{mail@larssonjohan.com} (0000-0002-4029-5945)

}
\keyword{internal}
