% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_presentation.R
\name{pdf_presentation}
\alias{pdf_presentation}
\title{PDF presentation template}
\usage{
pdf_presentation(..., template = NULL, page_numbering = TRUE,
  beamer_customization = TRUE)
}
\arguments{
\item{...}{arguments passed on to [rmarkdown::beamer_presentation()]}

\item{template}{what template to use. If `NULL`, the default
latex template will be used.}

\item{page_numbering}{whether to add page numbers to latex template}

\item{beamer_customization}{whether to add a block that allows
use of `\\setbeamertemplate\{\}[]`, `\\setbeamerfont\{\}[]`
through metadata blocks with analgous names and
with items `item` and `value` for setting various options.}
}
\value{
R Markdown output format to pass to [rmarkdown::render()]
}
\description{
PDF presentation template
}
