% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_presentation.R
\name{pdf_presentation}
\alias{pdf_presentation}
\title{PDF presentation template}
\usage{
pdf_presentation(
  ...,
  template = NULL,
  frame_numbering = TRUE,
  beamer_customization = TRUE,
  more_fontthemes = TRUE,
  subfigs = TRUE
)
}
\arguments{
\item{...}{arguments passed on to [rmarkdown::beamer_presentation()]}

\item{template}{what template to use. If `NULL`, the default
latex template will be used.}

\item{frame_numbering}{whether to add frame numbers to latex template}

\item{beamer_customization}{whether to add a block that allows
use of `\\setbeamertemplate\{\}[]`, `\\setbeamerfont\{\}[]`
through metadata blocks with analogous names and
with items `item` and `value` for setting various options.}

\item{more_fontthemes}{whether to add a block that
allows multiple fontthemes}

\item{subfigs}{whether to use the subfig package for subfig support}
}
\value{
R Markdown output format to pass to [rmarkdown::render()]
}
\description{
PDF presentation template
}
