% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdplugr.R
\docType{package}
\name{rmdplugr-package}
\alias{rmdplugr}
\alias{rmdplugr-package}
\title{rmdplugr: Plugins for R Markdown Formats}
\description{
Formats for R Markdown that undo modifications by
    'pandoc' and 'rmarkdown' to original 'latex' templates, such as
    smaller margins, paragraph spacing, and compact titles. In addition, 
    enhancements such as author blocks with affiliations and
    headers and footers are introduced. All of this functionality is built 
    around plugins that modify the default 'pandoc' template without relying on 
    custom templates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/rmdplugr}
  \item \url{https://jolars.github.io/rmdplugr/}
  \item Report bugs at \url{https://github.com/jolars/rmdplugr/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

}
\keyword{internal}
