% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdtemplates.R
\docType{package}
\name{rmdtemplates-package}
\alias{rmdtemplates}
\alias{rmdtemplates-package}
\title{rmdtemplates: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
